move adapters to avoid conflict with SDK in the future
