+++ plugins/zapp-analytics-plugin-gemius/android/src/main/java/com/applicaster/analytics/gemius/adapters/AnalyticsScreenAdapter.kt
@@ -1,4 +1,4 @@
-package com.applicaster.analytics.adapters
+package com.applicaster.analytics.gemius.adapters
 
 import androidx.annotation.CallSuper
 import com.applicaster.util.APLogger
